refactor: extract the mgmt and cmc to a shared crate
