+++ src/libs/shared/src/mgmt/mod.rs
@@ -1,0 +1,2 @@
+pub mod cmc;
+mod types;
