+++ src/mission_control/src/mgmt/mod.rs
@@ -1,3 +1,2 @@
-pub mod canister;
 pub mod constants;
 pub mod status;
